version 1

+++ backend/controllers/tareascontrollers.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler');
 const Tarea = require('../models/tareasModels')
 const getTareas = asyncHandler(async(req, res) => {
 
-    const tareas = await Tarea.find({})
+    const tareas = await Tarea.find({user: req.user.id})
     res.status(200).json(tareas);
     });
 
@@ -13,7 +13,8 @@ const createTareas = asyncHandler(async(req, res) => {
     }
 
     const tarea = await Tarea.create({
-        descripcion: req.body.descripcion
+        descripcion: req.body.descripcion,
+        user: req.user.id
     })
 
 res.status(201).json(tarea);
@@ -26,10 +27,16 @@ const updateTareas = asyncHandler(async(req, res) => {
         res.status(404)
         throw new Error("Esa tarea no existe")
     }
+    //nos aseguramos que el dueño de la tarea la pueda modificar
+    if(tarea.user.toString()!==req.user.id){
+        res.status(401)
+        throw new Error('usuario no autorzado')
+    }else{
+        const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
-    const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true})
-
-    res.status(200).json(tareaUpdated);
+        res.status(200).json(tareaUpdated);
+    }
+    
 });
 
 const deleteTareas = asyncHandler(async(req, res) => {
@@ -39,10 +46,16 @@ const deleteTareas = asyncHandler(async(req, res) => {
         res.status(404)
         throw new Error("Esa tarea no existe")
     }
+//solo el dueño la elimine 
+    if(tarea.user.toString()!== req.user.id){
+        res.status(401)
+        throw new Error('usuario no autorizado')
+    }else{
+        await Tarea.deleteOne(tarea)
 
-    await Tarea.deleteOne(tarea)
-
-    res.status(200).json({ id: req.params.id });
+        res.status(200).json({ id: req.params.id });
+    }
+    
 });
